Fixed lost overflowed lines in PDF, but do non use it for HTML, issue #1220
In PDF documentation on page 17 bottom are many lost overflowed lines of long code.
I divided it into two pages. Problem is, how to fix PDF and not to put wrong HTML.
I propose to apply this patch only if you will generate a PDF and then undo your changes in "settings.txt".